# api-docs/swagger.yaml
openapi: 3.0.3
info:
  title: CCTV Smart Attendance System API
  description: AI-powered attendance management with facial recognition, PPE detection, salary/fine management, and HR analytics.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login
  parameters:
    pageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    limitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    startDateParam:
      name: startDate
      in: query
      schema: { type: string, format: date }
    endDateParam:
      name: endDate
      in: query
      schema: { type: string, format: date }
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        code: { type: string }
    PaginatedResponse:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    User:
      type: object
      properties:
        _id: { type: string }
        employeeId: { type: string, example: EMP001 }
        firstName: { type: string, example: John }
        lastName: { type: string, example: Doe }
        email: { type: string, format: email }
        phone: { type: string }
        department: { type: string }
        designation: { type: string }
        baseSalary: { type: number, minimum: 0 }
        isActive: { type: boolean }
        role: { type: string, enum: [employee, admin, super-admin] }
        createdAt: { type: string, format: date-time }
...
    Fine:
      type: object
      properties:
        _id: { type: string }
        userId: { type: string }
        type:
          type: object
          properties:
            category: { type: string, enum: [late, absent, early-departure, safety-violation, ppe-violation, other] }
            ruleId: { type: string }
        amount: { type: number }
        currency: { type: string, default: USD }
        status: { type: string, enum: [pending, applied, disputed, waived] }
        evidence:
          type: object
          properties:
            description: { type: string }
            snapshotUrl: { type: string }
            timestamp: { type: string, format: date-time }
        dispute:
          type: object
          properties:
            reason: { type: string }
            filedAt: { type: string, format: date-time }
            resolvedAt: { type: string, format: date-time }
            resolution: { type: string }
...
    Loan:
      type: object
      properties:
        _id: { type: string }
        userId: { type: string }
        amount: { type: number }
        purpose: { type: string }
        totalInstallments: { type: integer }
        installmentAmount: { type: number }
        installmentsPaid: { type: integer }
        status: { type: string, enum: [active, completed, defaulted] }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
security:
  - bearerAuth: []
paths:
...
  /employees:
    get:
      tags: [Employees]
      summary: List employees (with search, filter, pagination)
      parameters:
        - { $ref: '#/components/parameters/pageParam' }
        - { $ref: '#/components/parameters/limitParam' }
        - { name: department, in: query, schema: { type: string } }
        - { name: isActive, in: query, schema: { type: boolean } }
        - { name: search, in: query, schema: { type: string }, description: Search by name or ID }
      responses:
        200:
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/PaginatedResponse' }
                  - type: object
                    properties:
                      employees: { type: array, items: { $ref: '#/components/schemas/User' } }
    post:
      tags: [Employees]
      summary: Create employee (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [employeeId, firstName, lastName, email, department]
              properties:
                employeeId: { type: string }
                firstName: { type: string }
                lastName: { type: string }
                email: { type: string, format: email }
                department: { type: string }
                baseSalary: { type: number, minimum: 0 }
                role: { type: string, enum: [employee, admin, super-admin] }
      responses:
        201:
          content:
            application/json:
              schema:
                type: object
                properties:
                  employee: { $ref: '#/components/schemas/User' }
  /employees/{id}:
    get:
      tags: [Employees]
      summary: Get employee by ID
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  employee: { $ref: '#/components/schemas/User' }
        404: { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Employees]
      summary: Update employee (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  employee: { $ref: '#/components/schemas/User' }
    delete:
      tags: [Employees]
      summary: Deactivate employee (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        200:
          description: Employee deactivated
  /employees/{id}/register-face:
    post:
      tags: [Employees]
      summary: Register face for recognition (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [imageBuffer]
              properties:
                imageBuffer: { type: string, description: Base64-encoded image (max 5MB) }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  confidence: { type: number }

  # ─── Attendance ────────────────────────────────────────────────
...
  /fines:
    get:
      tags: [Fines]
      summary: List fines (with filters)
      parameters:
        - { $ref: '#/components/parameters/pageParam' }
        - { $ref: '#/components/parameters/limitParam' }
        - { name: status, in: query, schema: { type: string, enum: [pending, applied, disputed, waived] } }
        - { $ref: '#/components/parameters/startDateParam' }
        - { $ref: '#/components/parameters/endDateParam' }
      responses:
        200:
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/PaginatedResponse' }
                  - type: object
                    properties:
                      fines: { type: array, items: { $ref: '#/components/schemas/Fine' } }
  /fines/user/{userId}:
    get:
      tags: [Fines]
      summary: Get fines for a specific user
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  fines: { type: array, items: { $ref: '#/components/schemas/Fine' } }
                  summary:
                    type: object
                    properties:
                      totalFines: { type: integer }
                      totalAmount: { type: number }
                      pending: { type: integer }
                      applied: { type: integer }
                      disputed: { type: integer }
                      waived: { type: integer }
  /fines/{id}/dispute:
    post:
      tags: [Fines]
      summary: Dispute a fine
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  fine: { $ref: '#/components/schemas/Fine' }
  /fines/{id}/resolve:
    post:
      tags: [Fines]
      summary: Resolve a fine dispute (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution]
              properties:
                resolution: { type: string }
                waive: { type: boolean, default: false }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  fine: { $ref: '#/components/schemas/Fine' }

  # ─── Fine Rules ────────────────────────────────────────────────
  /fine-rules:
    get:
      tags: [Fine Rules]
      summary: List fine rules
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules: { type: array, items: { type: object } }
    post:
      tags: [Fine Rules]
      summary: Create fine rule (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, category, amountType, amountValue]
              properties:
                name: { type: string }
                category: { type: string, enum: [late, absent, safety, ppe] }
                amountType: { type: string, enum: [fixed, percentage] }
                amountValue: { type: number, minimum: 0 }
                maxPerDay: { type: integer }
                maxPerMonth: { type: integer }
      responses:
        201: { description: Fine rule created }

  # ─── Salaries ──────────────────────────────────────────────────
...
  /loans:
    get:
      tags: [Loans]
      summary: List loans
      parameters:
        - { $ref: '#/components/parameters/pageParam' }
        - { $ref: '#/components/parameters/limitParam' }
        - { name: status, in: query, schema: { type: string, enum: [active, completed, defaulted] } }
      responses:
        200:
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/PaginatedResponse' }
                  - type: object
                    properties:
                      loans: { type: array, items: { $ref: '#/components/schemas/Loan' } }
  /loans/request:
    post:
      tags: [Loans]
      summary: Request a loan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, amount, purpose, installments]
              properties:
                userId: { type: string }
                amount: { type: number, minimum: 1 }
                purpose: { type: string }
                installments: { type: integer, minimum: 1, maximum: 12 }
      responses:
        201:
          content:
            application/json:
              schema:
                type: object
                properties:
                  loan: { $ref: '#/components/schemas/Loan' }
  /loans/user/{userId}:
    get:
      tags: [Loans]
      summary: Get loans for a user
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  loans: { type: array, items: { $ref: '#/components/schemas/Loan' } }
                  summary: { type: object }
  /loans/{id}/approve:
    post:
      tags: [Loans]
      summary: Approve a loan (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  loan: { $ref: '#/components/schemas/Loan' }
...
  /privacy/anonymize/{userId}:
    post:
      tags: [Privacy]
      summary: Anonymize user data — GDPR right to be forgotten (super-admin only)
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        200: { description: User data anonymized }
  /privacy/export/{userId}:
    get:
      tags: [Privacy]
      summary: Export user data — GDPR data portability
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        200: { description: User data export }
...
  /privacy/consent/{userId}:
    get:
      tags: [Privacy]
      summary: Get user consent status
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        200: { description: Consent data }
    put:
      tags: [Privacy]
      summary: Update user consent
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                faceRecognition: { type: boolean }
                emotionTracking: { type: boolean }
                locationTracking: { type: boolean }
                dataProcessing: { type: boolean }
      responses:
        200: { description: Consent updated }

  # ─── Monitoring ────────────────────────────────────────────────
...